Implement an IntersectableList class
Implement a class that contains a list of intersectable objects that is
itself Intersectable, where the intersect method will return the closest
object a given ray intersects.

Also add a method to the intersection struct for calculating the correct
normal at an intersection point depending on whether it hits the front
face of an object.