Refactor code
Refactor the code to help improve readability and future extensibility.
This involves moving files into relevant directories and restructuring
the tests to match. This also replaces `intersectables` with
`primatives` as this notion makes sense for objects with a texture (The
notion of a shape does not make sense because of what was previously
`IntersectableList`).